requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
groq>=0.5.0
openai>=1.24.1
requests>=2.31.0
numpy>=1.26.4
opencv-python-headless>=4.9.0.80
mediapipe>=0.10.0
